requirements.txt
aiofile==3.8.5
aiogram==2.25.1
aiohttp==3.8.4
aiohttpx==0.0.9
aiosignal==1.3.1
anyio==3.6.2
async-lru==2.0.2
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
backoff==2.2.1
caio==0.9.12
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
dill==0.3.6
exceptiongroup==1.1.1
file-io==0.4.5
frozendict==2.3.7
frozenlist==1.3.3
fsspec==2023.4.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
lazyops==0.2.37
loguru==0.7.0
magic-filter==1.0.9
multidict==6.0.4
openai==0.27.4
packaging==23.0
peewee==3.16.0
pluggy==1.0.0
pydantic==1.10.7
pytest==7.3.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
sniffio==1.3.0
tomli==2.0.1
tqdm==4.65.0
typing-extensions==4.5.0
urllib3==1.26.15
win32-setctime==1.1.0
yarl==1.8.2
